Fix #11570 - fix compile error
cast from uint32_t to int32_t

+++ src/openrct2-ui/windows/GameBottomToolbar.cpp
@@ -532,7 +532,7 @@ static void window_game_bottom_toolbar_draw_right_panel(rct_drawpixelinfo* dpi, 
 
     // Temperature
     screenCoords = { w->windowPos.x + window_game_bottom_toolbar_widgets[WIDX_RIGHT_OUTSET].left + 15,
-                     screenCoords.y + line_height + 1 };
+                     static_cast<int32_t>(screenCoords.y + line_height + 1) };
 
     int32_t temperature = gClimateCurrent.Temperature;
     rct_string_id format = STR_CELSIUS_VALUE;
